tests/libcxx/map

+++ test/tests/libcxx/map/map.access/iterator.pass.cpp
@@ -63,13 +63,13 @@ void main()
             V(8, 2)
         };
         map<int, double> m(ar, ar+sizeof(ar)/sizeof(ar[0]));
-        assert(std::distance(m.begin(), m.end()) == m.size());
-        assert(std::distance(m.rbegin(), m.rend()) == m.size());
+        assert((size_t)std::distance(m.begin(), m.end()) == m.size());
+        assert((size_t)std::distance(m.rbegin(), m.rend()) == m.size());
         map<int, double>::iterator i;
         i = m.begin();
         map<int, double>::const_iterator k = i;
         assert(i == k);
-        for (int j = 1; j <= m.size(); ++j, ++i)
+        for (int j = 1; j <= (int)m.size(); ++j, ++i)
         {
             assert(i->first == j);
             assert(i->second == 1);
@@ -107,13 +107,13 @@ void main()
             V(8, 2)
         };
         const map<int, double> m(ar, ar+sizeof(ar)/sizeof(ar[0]));
-        assert(std::distance(m.begin(), m.end()) == m.size());
-        assert(std::distance(m.cbegin(), m.cend()) == m.size());
-        assert(std::distance(m.rbegin(), m.rend()) == m.size());
-        assert(std::distance(m.crbegin(), m.crend()) == m.size());
+        assert((size_t)std::distance(m.begin(), m.end()) == m.size());
+        assert((size_t)std::distance(m.cbegin(), m.cend()) == m.size());
+        assert((size_t)std::distance(m.rbegin(), m.rend()) == m.size());
+        assert((size_t)std::distance(m.crbegin(), m.crend()) == m.size());
         map<int, double>::const_iterator i;
         i = m.begin();
-        for (int j = 1; j <= m.size(); ++j, ++i)
+        for (int j = 1; j <= (int)m.size(); ++j, ++i)
         {
             assert(i->first == j);
             assert(i->second == 1);
